+++ CHANGES.md
@@ -1,2 +1,12 @@
+* 0.2.3
+    * Templates now has functions
+        * to_upper
+        * to_lower
+        * to_str
+        * to_int
+        * round
+        * contains
+    * BUGFIX: component tick event happened after the widget cycle causing a
+      panic
 * 0.2.2
     * Everything: this is a complete rewrite